backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
google-generativeai==0.8.3
pydantic==2.10.4
pydantic-settings==2.7.1
Jinja2==3.1.5
aiosqlite==0.20.0
sqlalchemy[asyncio]==2.0.36
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
PyGithub==2.8.1
